+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/app/impl/base/InstalledAppStatusFetcher.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.util.Log
 import androidx.annotation.Keep
 import de.marmaro.krt.ffupdater.DisplayableException
-import de.marmaro.krt.ffupdater.FFUpdater
+import de.marmaro.krt.ffupdater.FFUpdater.Companion.LOG_TAG
 import de.marmaro.krt.ffupdater.app.VersionCompareHelper
 import de.marmaro.krt.ffupdater.app.entity.InstalledAppStatus
 import de.marmaro.krt.ffupdater.app.entity.LatestVersion
@@ -23,17 +23,17 @@ interface InstalledAppStatusFetcher : InstalledVersionFetcher, LatestVersionFetc
 
     suspend fun findInstalledAppStatus(context: Context, cacheBehaviour: CacheBehaviour): InstalledAppStatus {
         val available = try {
-            Log.d(FFUpdater.LOG_TAG, "findAppUpdateStatus(): Search for latest ${app.name} update.")
+            Log.d(LOG_TAG, "InstalledAppStatusFetcher: Search for latest ${app.name} update.")
             val (result, duration) = MeasureExecutionTime.measureMs {
                 fetchLatestUpdate(context.applicationContext, cacheBehaviour)
             }
-            Log.i(FFUpdater.LOG_TAG, "findAppUpdateStatus(): Found ${app.name} ${result.version} (${duration}ms).")
+            Log.i(LOG_TAG, "InstalledAppStatusFetcher: Found ${app.name} ${result.version} (${duration}ms).")
             result
         } catch (e: NetworkException) {
-            Log.d(FFUpdater.LOG_TAG, "findAppUpdateStatus(): Can't find latest update for ${app.name}.", e)
+            Log.d(LOG_TAG, "InstalledAppStatusFetcher: Can't find latest update for ${app.name}.", e)
             throw NetworkException("can't find latest update for ${app.name}.", e)
         } catch (e: DisplayableException) {
-            Log.d(FFUpdater.LOG_TAG, "findAppUpdateStatus(): Can't find latest update for ${app.name}.", e)
+            Log.d(LOG_TAG, "InstalledAppStatusFetcher: Can't find latest update for ${app.name}.", e)
             throw DisplayableException("can't find latest update for ${app.name}.", e)
         }
         return InstalledAppStatus(